+++ css/style.css
@@ -190,8 +190,8 @@ main {
   display: grid;
   grid-gap: 10px;
 
-  -ms-grid-columns: (1fr)[7];
-      grid-template-columns: repeat(7, 1fr);
+  /* -ms-grid-columns: (1fr)[7]; */
+  grid-template-columns: repeat(7, 1fr);
 }
 
 .picture {
@@ -1218,7 +1218,7 @@ main {
   top: 0;
   right: 0;
   bottom: 0;
-  z-index: 1;
+  z-index: 10;
 
   background-color: rgba(0, 0, 0, 0.8);
 }
@@ -1361,7 +1361,7 @@ main {
   }
 
   .pictures {
-    -ms-grid-columns: (1fr)[5];
+    /* -ms-grid-columns: (1fr)[5]; */
     grid-template-columns: repeat(5, 1fr);
   }
 
@@ -1382,7 +1382,7 @@ main {
   }
 
   .pictures {
-    -ms-grid-columns: (1fr)[3];
+    /* -ms-grid-columns: (1fr)[3]; */
     grid-template-columns: repeat(3, 1fr);
   }
 
@@ -1396,6 +1396,7 @@ main {
     grid-row: 1 / span 3;
   }
 }
+
 
 .visually-hidden {
   position: absolute;
