+++ index.html
@@ -9,7 +9,6 @@
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <script src="vendor/nouislider/nouislider.min.js"></script>
   <script src="vendor/pristine/pristine.min.js"></script>
-  <script src="js/functions.js"></script>
   <script type="module" src="js/main.js"></script>
   <title>Кекстаграм</title>
 </head>
@@ -159,7 +158,7 @@
           </div>
 
           <!-- Комментарии к изображению -->
-          <div class="social__comment-count">5 из <span class="comments-count">125</span> комментариев</div>
+          <div class="social__comment-count"><span class="social__comment-shown-count">5</span> из <span class="social__comment-total-count">125</span> комментариев</div>
           <ul class="social__comments">
             <li class="social__comment">
               <img class="social__picture" src="img/avatar-4.svg" alt="Аватар комментатора фотографии" width="35" height="35">
