+++ js/render-photo.js
@@ -1,0 +1,39 @@
+// Рендеринг миниатюр фотографий
+
+import { openBigPicture } from './big-picture.js';
+
+const picturesContainer = document.querySelector('.pictures');
+const pictureTemplate = document.querySelector('#picture').content.querySelector('.picture');
+
+const clearThumbnails = () => {
+  picturesContainer.querySelectorAll('.picture').forEach((picture) => picture.remove());
+};
+
+const renderThumbnail = (photo) => {
+  const pictureElement = pictureTemplate.cloneNode(true);
+  const pictureImg = pictureElement.querySelector('.picture__img');
+  const pictureComments = pictureElement.querySelector('.picture__comments');
+  const pictureLikes = pictureElement.querySelector('.picture__likes');
+
+  pictureImg.src = photo.url;
+  pictureImg.alt = photo.description;
+  pictureComments.textContent = photo.comments.length;
+  pictureLikes.textContent = photo.likes;
+
+  pictureElement.addEventListener('click', (evt) => {
+    evt.preventDefault();
+    openBigPicture(photo);
+  });
+
+  return pictureElement;
+};
+
+const renderThumbnails = (photos) => {
+  clearThumbnails();
+  const fragment = document.createDocumentFragment();
+  photos.forEach((photo) => fragment.appendChild(renderThumbnail(photo)));
+  picturesContainer.appendChild(fragment);
+};
+
+export { renderThumbnails };
+
